Move around types and refetch index
