+++ src/pages/library/datasetBuilder/constants.ts
@@ -1,4 +1,4 @@
-import { ConceptSet } from 'src/pages/library/datasetBuilder/dataset-builder-types';
+import { ConceptSet } from 'src/libs/ajax/DatasetBuilder';
 
 export const PAGE_PADDING_HEIGHT = 0;
 export const PAGE_PADDING_WIDTH = 3;
