+++ src/pages/library/datasetBuilder/dataset-builder-types.ts
@@ -1,45 +1,4 @@
-import {
-  DomainOption,
-  ProgramDataListOption,
-  ProgramDataListValueOption,
-  ProgramDataRangeOption,
-} from 'src/libs/ajax/DatasetBuilder';
-
-/** A specific criteria based on a type. */
-export interface Criteria {
-  kind: 'domain' | 'range' | 'list';
-  name: string;
-  id: number;
-  count: number;
-}
-
-export interface DomainCriteria extends Criteria {
-  kind: 'domain';
-  domainOption: DomainOption;
-}
-
-export interface ProgramDataRangeCriteria extends Criteria {
-  kind: 'range';
-  rangeOption: ProgramDataRangeOption;
-  low: number;
-  high: number;
-}
-export interface ProgramDataListCriteria extends Criteria {
-  kind: 'list';
-  listOption: ProgramDataListOption;
-  valuesSelected: ProgramDataListValueOption[];
-}
-
-export type AnyCriteria = DomainCriteria | ProgramDataRangeCriteria | ProgramDataListCriteria;
-
-/** A group of criteria. */
-export interface CriteriaGroup {
-  name: string;
-  criteria: Criteria[];
-  mustMeet: boolean;
-  meetAll: boolean;
-  count: number;
-}
+import { Cohort, CriteriaGroup, DomainOption } from 'src/libs/ajax/DatasetBuilder';
 
 let groupCount = 1;
 export const newCriteriaGroup = (): CriteriaGroup => {
@@ -52,24 +11,12 @@ export const newCriteriaGroup = (): CriteriaGroup => {
   };
 };
 
-export interface Cohort extends DatasetBuilderType {
-  criteriaGroups: CriteriaGroup[];
-}
-
 export const newCohort = (name: string): Cohort => {
   return {
     name,
     criteriaGroups: [],
   };
 };
-
-export interface ConceptSet extends DatasetBuilderType {
-  featureValueGroupName: string;
-}
-
-export interface DatasetBuilderType {
-  name: string;
-}
 
 type DatasetBuilderMode =
   | 'homepage'
