+++ src/pages/library/datasetBuilder/state.ts
@@ -1,6 +1,6 @@
+import { Cohort, ConceptSet } from 'src/libs/ajax/DatasetBuilder';
 import * as Utils from 'src/libs/utils';
 import { Atom } from 'src/libs/utils';
-import { Cohort, ConceptSet } from 'src/pages/library/datasetBuilder/dataset-builder-types';
 
 export const datasetBuilderCohorts: Atom<Cohort[]> = Utils.atom([] as Cohort[]);
